feat: components with modals refactored
