+++ src/components/ContactCard.tsx
@@ -1,15 +1,16 @@
-import { memo, useState } from 'react';
+import { memo } from 'react';
 import { ContactDto } from 'src/types/dto/ContactDto';
-import { Button, Card, Form, ListGroup, Modal } from 'react-bootstrap';
+import { Button, Card, ListGroup } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from 'src/redux/hooks';
 import {
   addContactToFavorite,
-  addContactToGroup,
   removeContact,
   removeContactFromFavorite,
   removeContactFromGroup,
 } from 'src/redux/contacts';
+import { useModal } from 'src/hooks/useModal';
+import { AddContactToGroupModal } from './AddContactToGroupModal';
 
 interface ContactCardProps {
   contact: ContactDto;
@@ -23,26 +24,18 @@ export const ContactCard = memo<ContactCardProps>(
     withLink,
     groupId,
   }) => {
-    const groupsState = useAppSelector(state => state.groups);
     const dispatch = useAppDispatch();
     const isFavorite = useAppSelector(state => state.favorite.includes(id));
 
-    const [show, setShow] = useState(false);
-
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-
-    const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-      e.preventDefault();
-      dispatch(
-        //@ts-ignore
-        addContactToGroup({ contactId: id, groupId: e.target[1].value })
-      );
-      handleClose();
-    };
+    const { show, handleClose, handleShow } = useModal();
 
     return (
       <>
+        <AddContactToGroupModal
+          handleClose={handleClose}
+          show={show}
+          contactId={id}
+        />
         <Card key={id}>
           <Card.Img variant='top' src={photo} />
           <Card.Body>
@@ -99,30 +92,6 @@ export const ContactCard = memo<ContactCardProps>(
             </Button>
           </Card.Body>
         </Card>
-        <Modal show={show} onHide={handleClose}>
-          <Form onSubmit={handleSubmit}>
-            <Modal.Header closeButton>
-              <Modal.Title>Choose group</Modal.Title>
-            </Modal.Header>
-            <Modal.Body>
-              <Form.Select aria-label='Default select example'>
-                {groupsState.map(group => (
-                  <option key={group.id} value={group.id}>
-                    {group.name}
-                  </option>
-                ))}
-              </Form.Select>
-            </Modal.Body>
-            <Modal.Footer>
-              <Button variant='secondary' onClick={handleClose}>
-                Close
-              </Button>
-              <Button variant='primary' type={'submit'}>
-                Add to selected group
-              </Button>
-            </Modal.Footer>
-          </Form>
-        </Modal>
       </>
     );
   }
